requirements.txt
farm-haystack[colab,inference]
streamlit
openai
requests